refactor: extract notes and read/search MCP tools into lithos.tools
Second extraction slice (10 more tools; 20 of 37 now live in lithos/tools/):

- tools/notes.py: lithos_write (~350 lines of boundary validation +
  outcome shaping), lithos_note_update, lithos_delete
- tools/read_search.py: lithos_read, lithos_search, lithos_retrieve,
  lithos_list, lithos_related, lithos_tags, lithos_node_stats — plus the
  relocated _bfs_provenance (now a module function taking the corpus
  manager), _CONTENT_QUERY_FTS_CAP, and _RELATED_INCLUDES, each moving
  with its sole consumers

Handler bodies are verbatim moves (self. -> server., span preamble ->
tool_span decorator + get_current_span). Envelope-conformance suite and
frozen tool snapshot pass unmodified. server.py shrinks to ~2,030 lines;
_register_tools now holds only the 17 task tools.

Part of the 2026-07 architecture review (tracker task 9f2be54f, PR 4 of 6).
