+++ src/lithos/server.py
@@ -1,15 +1,11 @@
 """Lithos MCP Server - FastMCP server exposing all tools."""
 
 import asyncio
-import collections
 import contextlib
-import dataclasses
-import hashlib
 import json
 import logging
-import math
 import time
-from datetime import UTC, datetime, timedelta
+from datetime import UTC, datetime
 from typing import Any
 
 from fastmcp import FastMCP
@@ -26,7 +22,7 @@ from lithos.envelopes import (
     error_envelope,
     invalid_input_envelope,
 )
-from lithos.errors import CoordinationError, SearchBackendError
+from lithos.errors import CoordinationError
 from lithos.events import (
     TASK_CANCELLED,
     TASK_CLAIMED,
@@ -39,16 +35,9 @@ from lithos.events import (
     LithosEvent,
 )
 from lithos.graph import KnowledgeGraph
-from lithos.intake import CorpusIntake, DeleteRequest, NoteUpdateRequest, WriteRequest
+from lithos.intake import CorpusIntake
 from lithos.knowledge import (
-    _UNSET,
-    VALID_ACCESS_SCOPES,
-    VALID_NOTE_TYPES,
-    VALID_STATUSES,
     KnowledgeManager,
-    _normalize_datetime,
-    _UnsetType,
-    validate_extra_metadata,
     validate_metadata_match,
 )
 from lithos.provenance import ProvenanceProjection
@@ -67,13 +56,6 @@ from lithos.tools import register_all
 from lithos.watch_intake import WatchIntake
 
 logger = logging.getLogger(__name__)
-
-
-# Practical upper bound for ``lithos_list(content_query=...)``. Tantivy
-# returns up to this many hits before Python-side filters run. A million
-# matches from a single FTS query would already be degenerate; at that
-# point the caller should tighten the query, not ask for more results.
-_CONTENT_QUERY_FTS_CAP = 1_000_000
 
 
 def _serialize_task_record(task: Task | TaskStatus) -> dict[str, Any]:
@@ -979,66 +961,6 @@ class LithosServer:
             # KnowledgeGraph._apply_reconcile flushes the cache itself
             # (graph.py:929-948); explicit save_cache() is redundant.
 
-    def _bfs_provenance(self, start_id: str, direction: str, depth: int) -> list[dict[str, str]]:
-        """BFS traversal over provenance indexes.
-
-        Args:
-            start_id: Starting document ID (excluded from results).
-            direction: "sources" or "derived".
-            depth: Maximum traversal depth (already clamped to 1-3).
-
-        Returns:
-            Sorted list of {id, title} dicts for discovered nodes.
-        """
-        visited: set[str] = {start_id}
-        frontier: collections.deque[str] = collections.deque()
-
-        # Seed the frontier with immediate neighbours
-        if direction == "sources":
-            for nid in self.knowledge.get_doc_sources(start_id):
-                if self.knowledge.has_document(nid) and nid not in visited:
-                    frontier.append(nid)
-                    visited.add(nid)
-        else:  # "derived"
-            for nid in self.knowledge.get_derived_docs(start_id):
-                if nid not in visited:
-                    frontier.append(nid)
-                    visited.add(nid)
-
-        current_depth = 1
-        result_ids: list[str] = list(frontier)
-
-        while current_depth < depth and frontier:
-            next_frontier: list[str] = []
-            for node_id in frontier:
-                if direction == "sources":
-                    neighbours = self.knowledge.get_doc_sources(node_id)
-                    for nid in neighbours:
-                        if self.knowledge.has_document(nid) and nid not in visited:
-                            next_frontier.append(nid)
-                            visited.add(nid)
-                else:
-                    neighbours = self.knowledge.get_derived_docs(node_id)
-                    for nid in neighbours:
-                        if nid not in visited:
-                            next_frontier.append(nid)
-                            visited.add(nid)
-            frontier = collections.deque(next_frontier)
-            result_ids.extend(next_frontier)
-            current_depth += 1
-
-        # Sort by ID for deterministic output, resolve titles
-        return sorted(
-            [
-                {
-                    "id": nid,
-                    "title": self.knowledge.get_title_by_id(nid),
-                }
-                for nid in result_ids
-            ],
-            key=lambda n: n["id"],
-        )
-
     async def stop_enrich_worker(self) -> None:
         """Stop the CognitiveMemory Module and close the projection.
 
@@ -1079,1249 +1001,6 @@ class LithosServer:
 
     def _register_tools(self) -> None:
         """Register all MCP tools."""
-
-        # ==================== Knowledge Tools ====================
-
-        @self.mcp.tool()
-        @tool_metrics()
-        async def lithos_write(
-            title: str,
-            content: str,
-            agent: str,
-            tags: list[str] | None = None,
-            confidence: float | None = None,
-            path: str | None = None,
-            id: str | None = None,
-            source_task: str | None = None,
-            source_url: str | None = None,
-            derived_from_ids: list[str] | None = None,
-            ttl_hours: float | None = None,
-            expires_at: str | None = None,
-            expected_version: int | None = None,
-            schema_version: int | None = None,
-            namespace: str | None = None,
-            access_scope: str | None = None,
-            note_type: str | None = None,
-            status: str | None = None,
-            summaries: dict | None = None,
-            metadata: dict | None = None,
-        ) -> dict[str, Any]:
-            """Create or update a knowledge file.
-
-            Args are grouped below by role. The grouping is documentation only —
-            all parameters remain flat at the MCP boundary. See
-            `docs/plans/unified-write-contract.md` for the normative field
-            contract.
-
-            Args:
-                title: Title of the knowledge item.
-                content: Markdown content (without frontmatter).
-                agent: Your agent identifier.
-
-                --- Identity & metadata ---
-                id: UUID to update existing; omit to create new.
-                tags: List of tags. On update: null/omit preserves existing; [] clears
-                    all tags; non-empty list replaces.
-                metadata: Free-form key/value dict persisted into the document's
-                    frontmatter. On update: null/omit preserves existing; {} clears
-                    all metadata; a non-empty dict is an additive per-key merge (a
-                    key whose value is null deletes it). Keys must be strings and
-                    must not collide with reserved frontmatter fields (e.g. title,
-                    tags, version) — such writes are rejected with code "invalid_input".
-                confidence: Confidence score 0-1 (default: 1.0 on create). On update:
-                    null/omit preserves existing; float sets new value. Integers
-                    are coerced to float; anything else that is not a finite
-                    number in [0.0, 1.0] (non-numeric, bool, NaN/inf, or
-                    out-of-range) is rejected with code "invalid_input".
-                path: Where to store the note. Two accepted forms:
-                    - Subdirectory (e.g., "procedures") — the filename is derived
-                      from the title (slugified) and ".md" is appended.
-                    - Full relative file path ending in ".md"
-                      (e.g., "procedures/my-doc.md") — the final segment is used
-                      as the filename verbatim; the title does NOT influence the
-                      filename in this mode.
-                    Intermediate path segments may not end in ".md"; such inputs
-                    are rejected with code "invalid_input" to prevent
-                    accidental creation of directories whose names end in ".md".
-
-                --- Provenance ---
-                source_url: URL provenance for this knowledge. On update: null/omit
-                    preserves existing; "" clears; string sets new value.
-                derived_from_ids: List of source document UUIDs this note was derived
-                    from. On update: null/omit preserves existing; [] clears;
-                    non-empty list replaces.
-                source_task: Task ID this knowledge came from.
-
-                --- Freshness ---
-                ttl_hours: Time-to-live in hours from now. Computes expires_at.
-                    Mutually exclusive with expires_at.
-                expires_at: Absolute ISO 8601 expiry datetime. On update: null/omit
-                    preserves existing; "" clears; ISO string sets new value.
-                    Mutually exclusive with ttl_hours.
-
-                --- Concurrency ---
-                expected_version: If provided on update, reject with version_conflict if the
-                    document's current version differs. Omit to skip version checking.
-                    On create, this parameter is silently ignored.
-
-                --- LCMA ---
-                schema_version: LCMA schema version (default 1 on create).
-                namespace: LCMA namespace. Persisted only if explicitly passed;
-                    derived at read time otherwise.
-                access_scope: shared|task|agent_private (default shared on create).
-                    task requires source_task.
-                note_type: observation|agent_finding|summary|concept|task_record|hypothesis
-                    (default observation on create).
-                status: active|archived|quarantined (default active on create).
-                summaries: Optional dict with short/long summary strings.
-
-            Returns:
-                On success: {"status": "created"|"updated", "id", "path", "version",
-                "warnings"}. Actionable outcomes keep their own top-level status:
-                duplicate / slug_collision / path_collision / version_conflict
-                (version_conflict carries "current_version"). All other failures
-                use the standard error envelope {"status": "error", "code",
-                "message"} — validation failures carry code "invalid_input".
-            """
-            logger.info("lithos_write agent=%s title=%r update=%s", agent, title, id is not None)
-            tracer = get_tracer()
-            with tracer.start_as_current_span("lithos.tool.write") as span:
-                span.set_attribute("lithos.tool", "lithos_write")
-                span.set_attribute("lithos.agent", agent)
-                span.set_attribute("lithos.is_update", id is not None)
-
-                # Validate ttl_hours / expires_at mutual exclusion
-                if ttl_hours is not None and expires_at is not None:
-                    return invalid_input_envelope(
-                        "Provide either ttl_hours or expires_at, not both."
-                    )
-
-                # Validate ttl_hours
-                if ttl_hours is not None and (
-                    not isinstance(ttl_hours, (int, float))
-                    or math.isnan(ttl_hours)
-                    or math.isinf(ttl_hours)
-                    or ttl_hours <= 0
-                ):
-                    return invalid_input_envelope("ttl_hours must be a finite positive number.")
-
-                # Validate LCMA enum fields
-                if access_scope is not None and access_scope not in VALID_ACCESS_SCOPES:
-                    return invalid_input_envelope(
-                        f"Invalid access_scope: {access_scope!r}. "
-                        f"Must be one of {sorted(VALID_ACCESS_SCOPES)}"
-                    )
-                if note_type is not None and note_type not in VALID_NOTE_TYPES:
-                    return invalid_input_envelope(
-                        f"Invalid note_type: {note_type!r}. "
-                        f"Must be one of {sorted(VALID_NOTE_TYPES)}"
-                    )
-                if status is not None and status not in VALID_STATUSES:
-                    return invalid_input_envelope(
-                        f"Invalid status: {status!r}. Must be one of {sorted(VALID_STATUSES)}"
-                    )
-
-                # Validate summaries shape
-                if summaries is not None:
-                    if not isinstance(summaries, dict):
-                        return invalid_input_envelope(
-                            "summaries must be an object with 'short' and/or 'long' string fields."
-                        )
-                    unknown_keys = set(summaries.keys()) - {"short", "long"}
-                    if unknown_keys:
-                        return invalid_input_envelope(
-                            f"summaries has unknown keys: {sorted(unknown_keys)}. "
-                            f"Allowed keys: ['long', 'short']."
-                        )
-                    for k, v in summaries.items():
-                        if not isinstance(v, str):
-                            return invalid_input_envelope(
-                                f"summaries.{k} must be a string, got {type(v).__name__}."
-                            )
-
-                # Validate metadata shape at the boundary for a fast, clean
-                # envelope. The same rule is enforced in the storage layer
-                # (KnowledgeManager) so the invariant holds for every caller.
-                if metadata is not None:
-                    try:
-                        validate_extra_metadata(metadata)
-                    except ValueError as e:
-                        return invalid_input_envelope(str(e))
-
-                # Validate task-scope create-time invariant. The update-time
-                # case is enforced under ``_write_lock`` inside
-                # ``KnowledgeManager.update`` to avoid the TOCTOU window
-                # described in ADR-0003.
-                if access_scope == "task" and id is None and not source_task:
-                    return invalid_input_envelope("access_scope='task' requires source_task")
-
-                # Emit freshness span attributes
-                if ttl_hours is not None:
-                    span.set_attribute("freshness.ttl_hours", ttl_hours)
-                elif expires_at is not None and expires_at != "":
-                    span.set_attribute("freshness.expires_at_set", True)
-
-                # Compute expires_at_dt from ttl_hours or expires_at string
-                expires_at_dt: datetime | None | _UnsetType
-                if ttl_hours is not None:
-                    expires_at_dt = datetime.now(UTC) + timedelta(hours=ttl_hours)
-                elif id is not None:
-                    # Update path: map MCP boundary to manager semantics
-                    # None (omitted) → _UNSET (preserve), "" → None (clear), str → parse
-                    if expires_at is None:
-                        expires_at_dt = _UNSET
-                    elif expires_at == "":
-                        expires_at_dt = None
-                    else:
-                        try:
-                            expires_at_dt = datetime.fromisoformat(expires_at)
-                            if expires_at_dt.tzinfo is None:
-                                expires_at_dt = expires_at_dt.replace(tzinfo=UTC)
-                            else:
-                                expires_at_dt = expires_at_dt.astimezone(UTC)
-                        except ValueError:
-                            return invalid_input_envelope(
-                                f"Invalid expires_at datetime: {expires_at}"
-                            )
-                else:
-                    # Create path: None means no expiry, str → parse
-                    if expires_at is None:
-                        expires_at_dt = None
-                    else:
-                        try:
-                            expires_at_dt = datetime.fromisoformat(expires_at)
-                            if expires_at_dt.tzinfo is None:
-                                expires_at_dt = expires_at_dt.replace(tzinfo=UTC)
-                            else:
-                                expires_at_dt = expires_at_dt.astimezone(UTC)
-                        except ValueError:
-                            return invalid_input_envelope(
-                                f"Invalid expires_at datetime: {expires_at}"
-                            )
-
-                # Translate MCP wire shape into intake field semantics:
-                #   None  (omitted) → _UNSET (preserve)
-                #   ""               → None (clear)
-                #   value            → set
-                # ``expires_at_dt`` already encodes this for the freshness
-                # field above; we mirror the same rule for the rest here.
-                if id is not None:
-                    url_arg: str | None | _UnsetType
-                    if source_url is None:
-                        url_arg = _UNSET
-                    elif source_url == "":
-                        url_arg = None
-                    else:
-                        url_arg = source_url
-                    prov_arg: list[str] | None | _UnsetType = (
-                        _UNSET if derived_from_ids is None else derived_from_ids
-                    )
-                    tags_arg: list[str] | _UnsetType = _UNSET if tags is None else tags
-                    conf_arg: float | _UnsetType = _UNSET if confidence is None else confidence
-                    source_arg: str | None | _UnsetType = (
-                        _UNSET if source_task is None else source_task
-                    )
-                    sv_arg: int | _UnsetType = _UNSET if schema_version is None else schema_version
-                    ns_arg: str | None | _UnsetType = _UNSET if namespace is None else namespace
-                    as_arg: str | None | _UnsetType = (
-                        _UNSET if access_scope is None else access_scope
-                    )
-                    nt_arg: str | None | _UnsetType = _UNSET if note_type is None else note_type
-                    st_arg: str | None | _UnsetType = _UNSET if status is None else status
-                    sum_arg: dict | None | _UnsetType = _UNSET if summaries is None else summaries
-                    meta_arg: dict | _UnsetType = _UNSET if metadata is None else metadata
-                else:
-                    # Create path forwards raw values; KnowledgeManager.create
-                    # applies its own defaults.
-                    url_arg = source_url or None
-                    prov_arg = derived_from_ids
-                    tags_arg = tags  # type: ignore[assignment]
-                    conf_arg = confidence  # type: ignore[assignment]
-                    source_arg = source_task
-                    sv_arg = schema_version  # type: ignore[assignment]
-                    ns_arg = namespace
-                    as_arg = access_scope
-                    nt_arg = note_type
-                    st_arg = status
-                    sum_arg = summaries
-                    meta_arg = metadata if metadata is not None else _UNSET
-
-                request = WriteRequest(
-                    title=title,
-                    content=content,
-                    id=id,
-                    tags=tags_arg,
-                    confidence=conf_arg,
-                    path=path,
-                    source_task=source_arg,
-                    source_url=url_arg,
-                    derived_from_ids=prov_arg,
-                    expires_at=expires_at_dt,
-                    expected_version=expected_version,
-                    schema_version=sv_arg,
-                    namespace=ns_arg,
-                    access_scope=as_arg,
-                    note_type=nt_arg,
-                    lcma_status=st_arg,
-                    summaries=sum_arg,
-                    metadata=meta_arg,
-                )
-
-                outcome = await self.intake.write(agent, request)
-
-                if outcome.status == "slug_collision":
-                    span.set_attribute("lithos.write_status", "slug_collision")
-                    return {
-                        "status": "slug_collision",
-                        "message": outcome.message,
-                        "existing_id": outcome.slug_collision_existing_id,
-                        "warnings": [],
-                    }
-                if outcome.status == "path_collision":
-                    span.set_attribute("lithos.write_status", "path_collision")
-                    return {
-                        "status": "path_collision",
-                        "message": outcome.message,
-                        "existing_id": outcome.path_collision_existing_id,
-                        "warnings": list(outcome.warnings),
-                    }
-                if outcome.status == "duplicate":
-                    span.set_attribute("lithos.write_status", "duplicate")
-                    dup = outcome.duplicate_of
-                    return {
-                        "status": "duplicate",
-                        "duplicate_of": {
-                            "id": dup.id,
-                            "title": dup.title,
-                            "source_url": dup.source_url,
-                        }
-                        if dup
-                        else None,
-                        "message": outcome.message,
-                        "warnings": list(outcome.warnings),
-                    }
-                if outcome.status == "version_conflict":
-                    # An actionable write outcome, not an error: read-merge-write
-                    # retry loops branch on this status (see CHANGELOG).
-                    span.set_attribute("lithos.write_status", "version_conflict")
-                    conflict: dict[str, Any] = {
-                        "status": "version_conflict",
-                        "message": outcome.message,
-                        "warnings": list(outcome.warnings),
-                    }
-                    if outcome.current_version is not None:
-                        conflict["current_version"] = outcome.current_version
-                    return conflict
-                if outcome.status in ("invalid_input", "content_too_large", "error"):
-                    span.set_attribute("lithos.write_status", outcome.status)
-                    if outcome.status == "invalid_input":
-                        return invalid_input_envelope(outcome.message or "")
-                    code = "internal_error" if outcome.status == "error" else outcome.status
-                    return error_envelope(code, outcome.message or "")
-
-                doc = outcome.document
-                assert doc is not None
-                span.set_attribute("lithos.doc_id", doc.id)
-                span.set_attribute("lithos.write_status", outcome.status)
-                span.set_attribute(
-                    "lithos.provenance.source_count",
-                    len(doc.metadata.derived_from_ids),
-                )
-                if outcome.warnings:
-                    span.set_attribute("lithos.provenance.warning_count", len(outcome.warnings))
-                logger.info("lithos_write completed doc_id=%s status=%s", doc.id, outcome.status)
-                return {
-                    "status": outcome.status,
-                    "id": doc.id,
-                    "path": str(doc.path),
-                    "version": doc.metadata.version,
-                    "warnings": list(outcome.warnings),
-                }
-
-        @self.mcp.tool()
-        @tool_metrics()
-        async def lithos_note_update(
-            id: str,
-            agent: str,
-            title: str | None = None,
-            tags: list[str] | None = None,
-            status: str | None = None,
-            metadata: dict | None = None,
-            expected_version: int | None = None,
-        ) -> dict[str, Any]:
-            """Patch a note's frontmatter (tags/metadata/title/status) without resending its body.
-
-            The note counterpart to ``lithos_task_update``: a per-field patch
-            that leaves the markdown body untouched. Use this instead of
-            ``lithos_write`` whenever you only need to change frontmatter — it
-            removes the read → reconstruct-body → write round-trip (and the
-            lost-update risk that comes with reproducing the body), since the
-            body is never read into the request at all.
-
-            At least one of ``title``, ``tags``, ``status``, or ``metadata`` must
-            be provided.
-
-            Args:
-                id: UUID of the note to patch.
-                agent: Your agent identifier.
-                title: New title. null/omit preserves existing. Renaming may
-                    change the note's slug; a collision with another note's
-                    slug is rejected as ``slug_collision``.
-                tags: null/omit preserves existing; ``[]`` clears all tags; a
-                    non-empty list replaces them.
-                status: active|archived|quarantined. null/omit preserves existing.
-                metadata: Additive per-key merge into the note's existing
-                    frontmatter metadata. A key whose value is null deletes it;
-                    other keys are set; keys not mentioned are preserved. As with
-                    ``lithos_task_update`` there is no wholesale-clear affordance —
-                    ``metadata={}`` makes no metadata change (and, with no other
-                    field set, is rejected with code "invalid_input"). Keys must be
-                    strings and must not collide with reserved frontmatter fields
-                    (e.g. title, tags, version) — such patches are rejected with
-                    code "invalid_input".
-                expected_version: If provided, reject with version_conflict when
-                    the note's current version differs. Omit to skip the check.
-
-            Returns:
-                On success: {"status": "updated", ...}. Actionable outcomes keep
-                their own top-level status: slug_collision / duplicate /
-                version_conflict (the latter carries "current_version"). All other
-                failures use the standard error envelope {"status": "error",
-                "code", "message"} — validation failures carry code
-                "invalid_input"; an unknown id carries code "note_not_found".
-            """
-            # An empty metadata dict carries no change (it maps to _UNSET below),
-            # so it does not count as a provided field — `metadata={}` alone is
-            # rejected rather than producing a version-bumping, event-emitting no-op.
-            has_metadata_change = metadata is not None and metadata != {}
-            if title is None and tags is None and status is None and not has_metadata_change:
-                return invalid_input_envelope(
-                    "At least one of title, tags, status, or metadata must be provided "
-                    "(an empty metadata dict makes no change)."
-                )
-
-            logger.info("lithos_note_update agent=%s id=%s", agent, id)
-            tracer = get_tracer()
-            with tracer.start_as_current_span("lithos.tool.note_update") as span:
-                span.set_attribute("lithos.tool", "lithos_note_update")
-                span.set_attribute("lithos.agent", agent)
-                span.set_attribute("lithos.doc_id", id)
-
-                # Validate status enum at the boundary for a fast, clean envelope.
-                if status is not None and status not in VALID_STATUSES:
-                    return invalid_input_envelope(
-                        f"Invalid status: {status!r}. Must be one of {sorted(VALID_STATUSES)}"
-                    )
-
-                # Validate metadata shape at the boundary. The same rule is
-                # enforced in KnowledgeManager so the invariant holds for every
-                # caller; checking here gives a fast, clean envelope.
-                if metadata is not None:
-                    try:
-                        validate_extra_metadata(metadata)
-                    except ValueError as e:
-                        return invalid_input_envelope(str(e))
-
-                # Translate the MCP wire shape into KnowledgeManager semantics:
-                #   None (omitted) → _UNSET (preserve)
-                #   tags=[]        → clear all tags
-                #   metadata={}    → _UNSET (no-op, mirroring lithos_task_update)
-                tags_arg: list[str] | _UnsetType = _UNSET if tags is None else tags
-                st_arg: str | None | _UnsetType = _UNSET if status is None else status
-                meta_arg: dict | _UnsetType = (
-                    _UNSET if metadata is None or metadata == {} else metadata
-                )
-
-                request = NoteUpdateRequest(
-                    id=id,
-                    title=title,
-                    tags=tags_arg,
-                    lcma_status=st_arg,
-                    metadata=meta_arg,
-                    expected_version=expected_version,
-                )
-
-                try:
-                    outcome = await self.intake.note_update(agent, request)
-                except FileNotFoundError:
-                    span.set_attribute("lithos.write_status", "note_not_found")
-                    return error_envelope("note_not_found", f"Note {id} not found")
-
-                if outcome.status == "slug_collision":
-                    span.set_attribute("lithos.write_status", "slug_collision")
-                    return {
-                        "status": "slug_collision",
-                        "message": outcome.message,
-                        "existing_id": outcome.slug_collision_existing_id,
-                        "warnings": [],
-                    }
-                if outcome.status == "duplicate":
-                    span.set_attribute("lithos.write_status", "duplicate")
-                    dup = outcome.duplicate_of
-                    return {
-                        "status": "duplicate",
-                        "duplicate_of": {
-                            "id": dup.id,
-                            "title": dup.title,
-                            "source_url": dup.source_url,
-                        }
-                        if dup
-                        else None,
-                        "message": outcome.message,
-                        "warnings": list(outcome.warnings),
-                    }
-                if outcome.status == "version_conflict":
-                    # An actionable write outcome, not an error: read-merge-write
-                    # retry loops branch on this status (see CHANGELOG).
-                    span.set_attribute("lithos.write_status", "version_conflict")
-                    conflict: dict[str, Any] = {
-                        "status": "version_conflict",
-                        "message": outcome.message,
-                        "warnings": list(outcome.warnings),
-                    }
-                    if outcome.current_version is not None:
-                        conflict["current_version"] = outcome.current_version
-                    return conflict
-                if outcome.status in ("invalid_input", "error"):
-                    span.set_attribute("lithos.write_status", outcome.status)
-                    if outcome.status == "invalid_input":
-                        return invalid_input_envelope(outcome.message or "")
-                    return error_envelope("internal_error", outcome.message or "")
-
-                doc = outcome.document
-                assert doc is not None
-                span.set_attribute("lithos.write_status", outcome.status)
-                logger.info(
-                    "lithos_note_update completed doc_id=%s status=%s", doc.id, outcome.status
-                )
-                return {
-                    "status": outcome.status,
-                    "id": doc.id,
-                    "path": str(doc.path),
-                    "version": doc.metadata.version,
-                    "warnings": list(outcome.warnings),
-                }
-
-        @self.mcp.tool()
-        @tool_metrics()
-        async def lithos_read(
-            id: str | None = None,
-            path: str | None = None,
-            max_length: int | None = None,
-            agent_id: str | None = None,
-        ) -> dict[str, Any]:
-            """Read a knowledge file by ID or path.
-
-            Args:
-                id: UUID of knowledge item
-                path: File path relative to knowledge/
-                max_length: Truncate content to N characters
-                agent_id: Caller identity for audit logging (optional)
-
-            Returns:
-                Dict with id, title, content, metadata, links, truncated,
-                retrieval_count
-            """
-            logger.info("lithos_read id=%s path=%s", id, path)
-            tracer = get_tracer()
-            with tracer.start_as_current_span("lithos.tool.read") as span:
-                span.set_attribute("lithos.tool", "lithos_read")
-                if id:
-                    span.set_attribute("lithos.id", id)
-
-                try:
-                    doc, truncated = await self.knowledge.read(
-                        id=id,
-                        path=path,
-                        max_length=max_length,
-                    )
-                except FileNotFoundError as e:
-                    return error_envelope("doc_not_found", str(e))
-
-                # Audit log — awaited so the write is committed before we query
-                # retrieval_count (avoids TOCTOU off-by-one). lithos_search uses
-                # fire-and-forget (asyncio.create_task) for its batch write since
-                # retrieval_count accuracy is not required there.
-                audit_agent = agent_id or "unknown"
-                await self.coordination.log_access(
-                    doc_id=doc.id,
-                    operation="read",
-                    agent_id=audit_agent,
-                )
-
-                # Retrieval count — how many times this doc has been read
-                retrieval_count = await self.coordination.get_retrieval_count(doc.id)
-
-                span.set_attribute("lithos.truncated", truncated)
-                meta = doc.metadata.to_dict()
-                meta["source_url"] = doc.metadata.source_url  # null when None
-                meta.setdefault("derived_from_ids", [])
-                # Free-form key/value metadata (#305) as an isolated dict, so
-                # callers can read back exactly what they wrote via the
-                # lithos_write `metadata` param without sifting reserved fields.
-                meta["extra"] = dict(doc.metadata.extra)
-                return {
-                    "id": doc.id,
-                    "title": doc.title,
-                    "content": doc.content,
-                    "metadata": meta,
-                    "links": [
-                        {"target": link.target, "display": link.display} for link in doc.links
-                    ],
-                    "truncated": truncated,
-                    "retrieval_count": retrieval_count,
-                }
-
-        @self.mcp.tool()
-        @tool_metrics()
-        async def lithos_delete(
-            id: str,
-            agent: str,
-        ) -> dict[str, Any]:
-            """Delete a knowledge file.
-
-            Args:
-                id: UUID of knowledge item to delete
-                agent: Agent performing deletion (required for audit trail)
-
-            Returns:
-                Dict with success boolean, or error envelope if document not found
-            """
-            logger.info("lithos_delete id=%s agent=%s", id, agent)
-            tracer = get_tracer()
-            with tracer.start_as_current_span("lithos.tool.delete") as span:
-                span.set_attribute("lithos.tool", "lithos_delete")
-                span.set_attribute("lithos.id", id)
-                span.set_attribute("lithos.agent", agent)
-
-                outcome = await self.intake.delete(agent, DeleteRequest(id=id))
-                if outcome.status == "not_found":
-                    return error_envelope("doc_not_found", f"Document not found: {id}")
-                return {"success": True}
-
-        @self.mcp.tool()
-        @tool_metrics()
-        async def lithos_search(
-            query: str,
-            limit: int = 10,
-            mode: str = "hybrid",
-            tags: list[str] | None = None,
-            author: str | None = None,
-            path_prefix: str | None = None,
-            threshold: float | None = None,
-            seed_ids: list[str] | None = None,
-            graph_depth: int = 2,
-            entities: list[str] | None = None,
-            agent_id: str | None = None,
-        ) -> dict[str, Any]:
-            """Search across the knowledge base.
-
-            Supports four search modes:
-            - ``hybrid`` (default): Merges Tantivy BM25 full-text and ChromaDB
-              cosine-similarity results using Reciprocal Rank Fusion (RRF, k=60).
-              Best overall quality.
-            - ``fulltext``: Pure Tantivy full-text search (BM25). Supports Tantivy
-              query syntax (e.g. field-specific queries, boolean operators).
-            - ``semantic``: Pure ChromaDB semantic/vector search. Finds documents
-              with similar meaning even when keywords differ.
-            - ``graph``: Wiki-link graph traversal. Discovers related documents by
-              following links from seed documents up to *graph_depth* hops.
-              Seeds are either provided via *seed_ids* or discovered automatically
-              via a fast hybrid search on *query*.
-
-            Args:
-                query: Search query string
-                limit: Max results (default: 10)
-                mode: Search mode — "hybrid" | "fulltext" | "semantic" | "graph"
-                      (default: "hybrid")
-                tags: Filter by tags (AND) — fulltext/semantic/hybrid only
-                author: Filter by author (fulltext/semantic/hybrid only)
-                path_prefix: Filter by path prefix (fulltext/semantic/hybrid only)
-                threshold: Minimum similarity 0-1 for semantic/hybrid (default: 0.5)
-                seed_ids: Starting document IDs for graph mode.  If omitted,
-                          seeds are discovered via hybrid search.
-                graph_depth: BFS hop depth for graph mode (1-3, default: 2)
-                entities: Filter results to documents whose ``entities``
-                          frontmatter contains every named entity (exact
-                          match, AND). Applies to all modes; resolved via an
-                          inverted index and applied as a post-filter.
-                agent_id: Caller identity for audit logging (optional)
-
-            Returns:
-                Dict with results list containing id, title, snippet, score, path,
-                source_url, updated_at, is_stale, derived_from_ids
-            """
-            logger.info("lithos_search mode=%s query_len=%d limit=%d", mode, len(query), limit)
-            tracer = get_tracer()
-            with tracer.start_as_current_span("lithos.tool.search") as span:
-                span.set_attribute("lithos.tool", f"lithos_search:{mode}")
-                span.set_attribute("lithos.query.length", len(query))
-                span.set_attribute(
-                    "lithos.query.sha256",
-                    hashlib.sha256(query.encode()).hexdigest()[:16],
-                )
-                span.set_attribute("lithos.limit", limit)
-                span.set_attribute("lithos.mode", mode)
-
-                valid_modes = {"hybrid", "fulltext", "semantic", "graph"}
-                if mode not in valid_modes:
-                    return error_envelope(
-                        "invalid_mode",
-                        f"Unknown search mode {mode!r}. Valid values: hybrid, fulltext, semantic, graph.",
-                    )
-
-                # Entities are not a per-backend filter: resolve the candidate
-                # set once from the knowledge inverted index (#316) and
-                # post-filter every mode's hits. Over-fetch to compensate,
-                # mirroring the engine's own post-filter heuristic.
-                entity_candidates = self.knowledge.entities_candidate_ids(entities)
-                if entity_candidates is not None and not entity_candidates:
-                    # No document carries every requested entity — skip the
-                    # backend search entirely.
-                    return {"results": []}
-                fetch_limit = limit * 5 if entity_candidates is not None else limit
-
-                def _build_result(r: Any, score_attr: str = "score") -> dict[str, Any]:
-                    return {
-                        "id": r.id,
-                        "title": r.title,
-                        "snippet": r.snippet,
-                        "score": getattr(r, score_attr),
-                        "path": r.path,
-                        "source_url": r.source_url,
-                        "updated_at": r.updated_at,
-                        "is_stale": r.is_stale,
-                        "derived_from_ids": self.knowledge.get_doc_sources(r.id),
-                    }
-
-                # Thread safety note: SearchManager read methods (full_text_search, semantic_search,
-                # hybrid_search) and the mutating methods (index, remove) are
-                # all wrapped in asyncio.to_thread() so Tantivy commits and ChromaDB embedding
-                # don't block the event loop. Concurrent read+write is not protected by a lock,
-                # but tantivy-py and ChromaDB are thread-safe for these operations. The
-                # embedding model is loaded eagerly in SearchEngine.create() at server
-                # startup, so no model-init race is reachable here.
-                if mode == "fulltext":
-                    ft_results = await asyncio.to_thread(
-                        self.search.full_text_search,
-                        query=query,
-                        limit=fetch_limit,
-                        tags=tags,
-                        author=author,
-                        path_prefix=path_prefix,
-                    )
-                    results_payload = [_build_result(r) for r in ft_results]
-                elif mode == "semantic":
-                    sem_results = await asyncio.to_thread(
-                        self.search.semantic_search,
-                        query=query,
-                        limit=fetch_limit,
-                        threshold=threshold,
-                        tags=tags,
-                        author=author,
-                        path_prefix=path_prefix,
-                    )
-                    results_payload = [
-                        _build_result(r, score_attr="similarity") for r in sem_results
-                    ]
-                elif mode == "graph":
-                    graph_results = await asyncio.to_thread(
-                        self.search.graph_search,
-                        query=query,
-                        graph=self.graph,
-                        seed_ids=seed_ids,
-                        depth=graph_depth,
-                        limit=fetch_limit,
-                        tags=tags,
-                        author=author,
-                        path_prefix=path_prefix,
-                        threshold=threshold,
-                    )
-                    results_payload = [_build_result(r) for r in graph_results]
-                else:
-                    # hybrid (default)
-                    hybrid_results = await asyncio.to_thread(
-                        self.search.hybrid_search,
-                        query=query,
-                        limit=fetch_limit,
-                        threshold=threshold,
-                        tags=tags,
-                        author=author,
-                        path_prefix=path_prefix,
-                    )
-                    results_payload = [_build_result(r) for r in hybrid_results]
-
-                if entity_candidates is not None:
-                    results_payload = [r for r in results_payload if r["id"] in entity_candidates][
-                        :limit
-                    ]
-
-                span.set_attribute("lithos.result_count", len(results_payload))
-                logger.info("lithos_search mode=%s results=%d", mode, len(results_payload))
-
-                # Audit log every returned document in a single batch write — fire-and-forget.
-                # Using log_access_batch avoids N concurrent SQLite connections (previously
-                # one asyncio.create_task per result document).
-                audit_agent = agent_id or "unknown"
-                asyncio.create_task(  # noqa: RUF006
-                    self.coordination.log_access_batch(
-                        doc_ids=[r["id"] for r in results_payload],
-                        operation="search_result",
-                        agent_id=audit_agent,
-                    )
-                )
-
-                return {"results": results_payload}
-
-        # ==================== LCMA Retrieval ====================
-
-        @self.mcp.tool()
-        @tool_metrics()
-        async def lithos_retrieve(
-            query: str,
-            limit: int = 10,
-            namespace_filter: list[str] | None = None,
-            agent_id: str | None = None,
-            task_id: str | None = None,
-            surface_conflicts: bool = False,
-            max_context_nodes: int | None = None,
-            tags: list[str] | None = None,
-            path_prefix: str | None = None,
-        ) -> dict[str, Any]:
-            """LCMA cognitive retrieval — runs seven scouts with reranking.
-
-            Orchestrates parallel scouts against the knowledge base, applies
-            merge-and-normalize, Terrace 1 reranking, and writes an audit
-            receipt on every call.
-
-            Args:
-                query: Search query string (required)
-                limit: Max results (default: 10)
-                namespace_filter: Restrict to these namespaces
-                agent_id: Caller identity for access-scope gating and audit
-                task_id: Task context — activates task_context scout and
-                    working-memory tracking
-                surface_conflicts: Reserved for MVP 2 contradiction surfacing
-                max_context_nodes: Provenance seed count (defaults to limit)
-                tags: Filter by tags (AND semantics)
-                path_prefix: Filter by path prefix
-
-            Returns:
-                Dict with results list (superset of lithos_search result
-                schema), temperature, terrace_reached, and receipt_id.
-            """
-            logger.info(
-                "lithos_retrieve: called",
-                extra={
-                    "query_len": len(query),
-                    "limit": limit,
-                    "agent_id": agent_id,
-                    "task_id": task_id,
-                    "namespace_filter": namespace_filter,
-                    "surface_conflicts": surface_conflicts,
-                },
-            )
-            tracer = get_tracer()
-            with tracer.start_as_current_span("lithos.tool.retrieve") as span:
-                span.set_attribute("lithos.tool", "lithos_retrieve")
-                span.set_attribute("lithos.query.length", len(query))
-                span.set_attribute("lithos.limit", limit)
-
-                # Envelope-shaping short-circuit: when LCMA is disabled the
-                # CognitiveMemory.retrieve precondition would fail, so we
-                # surface a typed error response instead. The Module method
-                # itself stays free of envelope concerns.
-                if not self._config.lcma.enabled:
-                    logger.warning("lithos_retrieve: LCMA is disabled")
-                    return error_envelope("lcma_disabled", "LCMA is disabled via configuration")
-
-                result = await self.memory.retrieve(
-                    query=query,
-                    limit=limit,
-                    namespace_filter=namespace_filter,
-                    agent_id=agent_id,
-                    task_id=task_id,
-                    surface_conflicts=surface_conflicts,
-                    max_context_nodes=max_context_nodes,
-                    tags=tags,
-                    path_prefix=path_prefix,
-                )
-
-                result_count = len(result.get("results", []))  # type: ignore[union-attr]
-                span.set_attribute("lithos.result_count", result_count)
-                logger.info(
-                    "lithos_retrieve: completed",
-                    extra={
-                        "result_count": result_count,
-                        "receipt_id": result.get("receipt_id"),  # type: ignore[union-attr]
-                        "temperature": result.get("temperature"),  # type: ignore[union-attr]
-                        "terrace_reached": result.get("terrace_reached"),  # type: ignore[union-attr]
-                        "agent_id": agent_id,
-                        "task_id": task_id,
-                    },
-                )
-                return result  # type: ignore[return-value]
-
-        @self.mcp.tool()
-        @tool_metrics()
-        async def lithos_list(
-            path_prefix: str | None = None,
-            tags: list[str] | None = None,
-            author: str | None = None,
-            since: str | None = None,
-            limit: int = 50,
-            offset: int = 0,
-            title_contains: str | None = None,
-            content_query: str | None = None,
-            metadata_match: dict | None = None,
-            entities: list[str] | None = None,
-        ) -> dict[str, Any]:
-            """List knowledge documents with filters.
-
-            Args:
-                path_prefix: Filter by path prefix
-                tags: Filter by tags (AND)
-                author: Filter by author
-                since: Filter by updated since (ISO datetime)
-                limit: Max results (default: 50)
-                offset: Pagination offset
-                title_contains: Filter by case-insensitive substring match on title
-                entities: Filter by entity names from the document's ``entities``
-                    frontmatter (AND across the list, exact match). Resolved via
-                    an inverted index (no full scan).
-                metadata_match: Filter by free-form metadata (AND across keys). For
-                    each ``key: q`` a document matches when its stored metadata
-                    value equals ``q`` or is a list containing ``q`` (so a note with
-                    ``github_repos: ["org/a","org/b"]`` matches
-                    ``{"github_repos": "org/a"}``). Query values must be scalars
-                    (string/number/boolean); type-sensitive. Resolved via an
-                    inverted index (no full scan).
-                content_query: Filter by full-text search query (Tantivy).
-                    Tantivy-native filters (``tags``, ``author``,
-                    ``path_prefix``) are pushed down into the search query so
-                    ranking runs over the already-filtered candidate set,
-                    which is necessary for correctness under ranking pressure
-                    (see #194). ``since`` and ``title_contains`` are applied
-                    against the metadata cache after ranking.
-
-            Returns:
-                Dict with items list and total count
-            """
-            logger.info("lithos_list limit=%d offset=%d", limit, offset)
-            tracer = get_tracer()
-            with tracer.start_as_current_span("lithos.tool.list") as span:
-                span.set_attribute("lithos.tool", "lithos_list")
-                span.set_attribute("lithos.limit", limit)
-
-                # Normalize to UTC so comparisons against normalized document
-                # timestamps never mix naive and aware values.
-                since_dt = None
-                if since:
-                    try:
-                        since_dt = _normalize_datetime(datetime.fromisoformat(since))
-                    except ValueError:
-                        return invalid_input_envelope(f"Invalid since datetime: {since}")
-
-                if metadata_match is not None:
-                    try:
-                        validate_metadata_match(metadata_match)
-                    except ValueError as e:
-                        return invalid_input_envelope(str(e))
-
-                if content_query is not None:
-                    # Push the Tantivy-native filters into the search call so
-                    # ranking runs over the filtered candidate set. Using a
-                    # global ranked window and then intersecting (the prior
-                    # approach) silently dropped matches when filtered docs
-                    # ranked deep globally — see #194.
-                    try:
-                        fts_results = await asyncio.to_thread(
-                            self.search.full_text_search,
-                            query=content_query,
-                            limit=_CONTENT_QUERY_FTS_CAP,
-                            tags=tags,
-                            author=author,
-                            path_prefix=path_prefix,
-                        )
-                    except SearchBackendError as exc:
-                        return error_envelope(
-                            "search_backend_error", f"Full-text search failed: {exc}"
-                        )
-
-                    # Apply the filters Tantivy doesn't handle: ``since``,
-                    # ``title_contains`` and ``metadata_match``. Consult the
-                    # metadata cache / inverted index so we don't incur a disk
-                    # read per candidate. ``meta_candidates`` is None when no
-                    # metadata filter is requested.
-                    #
-                    # Like ``since``/``title_contains``, ``metadata_match`` is a
-                    # post-rank filter here: metadata isn't a Tantivy field, so
-                    # it can't be pushed into the query the way tags/author/
-                    # path_prefix are (#194). It runs against the full ranked
-                    # window (cap = _CONTENT_QUERY_FTS_CAP = 1e6), so a match is
-                    # only ever dropped if >1e6 docs match content_query — the
-                    # same bound the other two post-filters already accept.
-                    meta_candidates = self.knowledge.metadata_candidate_ids(metadata_match)
-                    entity_candidates = self.knowledge.entities_candidate_ids(entities)
-                    if (meta_candidates is not None and not meta_candidates) or (
-                        entity_candidates is not None and not entity_candidates
-                    ):
-                        # An equality filter matched nothing — no point walking
-                        # the ranked window.
-                        fts_results = []
-                    matching_ids: list[str] = []
-                    for r in fts_results:
-                        if meta_candidates is not None and r.id not in meta_candidates:
-                            continue
-                        if entity_candidates is not None and r.id not in entity_candidates:
-                            continue
-                        cached = self.knowledge.get_cached_meta(r.id)
-                        if cached is None:
-                            continue
-                        if since_dt is not None:
-                            cached_updated = _normalize_datetime(cached.updated_at)
-                            if cached_updated < since_dt:
-                                continue
-                        if (
-                            title_contains is not None
-                            and title_contains.lower() not in (cached.title or "").lower()
-                        ):
-                            continue
-                        matching_ids.append(r.id)
-
-                    total = len(matching_ids)
-                    page_ids = matching_ids[offset : offset + limit]
-                    docs = []
-                    for doc_id in page_ids:
-                        try:
-                            doc, _ = await self.knowledge.read(id=doc_id)
-                            docs.append(doc)
-                        except Exception:
-                            # Cache can briefly lag the filesystem during
-                            # reconcile; skipping mirrors list_all's behaviour.
-                            continue
-                elif title_contains is not None:
-                    # ``title_contains`` has no Tantivy-backed fast path; fall
-                    # back to list_all and filter in memory. Tracked in #201
-                    # for a metadata-cache-only variant.
-                    _, total_base = await self.knowledge.list_all(
-                        path_prefix=path_prefix,
-                        tags=tags,
-                        author=author,
-                        since=since_dt,
-                        metadata_match=metadata_match,
-                        entities=entities,
-                        limit=0,
-                        offset=0,
-                    )
-                    if total_base > 0:
-                        all_docs, _ = await self.knowledge.list_all(
-                            path_prefix=path_prefix,
-                            tags=tags,
-                            author=author,
-                            since=since_dt,
-                            metadata_match=metadata_match,
-                            entities=entities,
-                            limit=total_base,
-                            offset=0,
-                        )
-                    else:
-                        all_docs = []
-                    all_docs = [d for d in all_docs if title_contains.lower() in d.title.lower()]
-                    total = len(all_docs)
-                    docs = all_docs[offset : offset + limit]
-                else:
-                    docs, total = await self.knowledge.list_all(
-                        path_prefix=path_prefix,
-                        tags=tags,
-                        author=author,
-                        since=since_dt,
-                        metadata_match=metadata_match,
-                        entities=entities,
-                        limit=limit,
-                        offset=offset,
-                    )
-
-                span.set_attribute("lithos.result_count", len(docs))
-                logger.info("lithos_list results=%d total=%d", len(docs), total)
-                return {
-                    "items": [
-                        {
-                            "id": d.id,
-                            "title": d.title,
-                            "path": str(d.path),
-                            "updated": d.metadata.updated_at.isoformat(),
-                            "tags": d.metadata.tags,
-                            "source_url": d.metadata.source_url or "",
-                            "derived_from_ids": self.knowledge.get_doc_sources(d.id),
-                            "metadata": dict(d.metadata.extra),
-                        }
-                        for d in docs
-                    ],
-                    "total": total,
-                }
-
-        # ==================== Graph Tools ====================
-
-        _RELATED_INCLUDES = ("links", "provenance", "edges")
-
-        @self.mcp.tool()
-        @tool_metrics()
-        async def lithos_related(
-            id: str,
-            include: list[str] | None = None,
-            depth: int = 1,
-            namespace: str | None = None,
-        ) -> dict[str, Any]:
-            """Composite "what is this document related to?" view.
-
-            Merges three graph-query backends into a single response so agents
-            don't have to fan out across multiple tools and mentally join the
-            results:
-
-            - **links** — structural ``[[wiki-link]]`` navigation (NetworkX).
-            - **provenance** — ``derived_from_ids`` chains (frontmatter index).
-            - **edges** — typed LCMA edges (edges.db), both directions.
-
-            For edge-table queries that are not centred on a single document
-            (e.g. "list all ``contradicts`` edges", "audit a namespace"), use
-            :func:`lithos_edge_list` instead — that tool is the only way to
-            express filters like ``type`` alone or ``to_id`` alone.
-
-            Args:
-                id: Document UUID.
-                include: Subset of ``["links", "provenance", "edges"]`` to
-                    populate. Defaults to all three. Unknown values are
-                    silently ignored so forward-compatible callers don't
-                    break when new backends land.
-                depth: BFS depth 1-3 for ``links`` and ``provenance``.
-                    Ignored for ``edges`` (LCMA edges are a flat table).
-                namespace: Optional namespace filter applied to ``edges``.
-                    ``links`` and ``provenance`` don't use namespaces.
-
-            Returns:
-                Dict shaped like::
-
-                    {
-                      "id": "<doc-id>",
-                      "included": ["links", "provenance", "edges"],
-                      "links": {"outgoing": [...], "incoming": [...]},
-                      "provenance": {
-                          "sources": [...], "derived": [...],
-                          "unresolved_sources": [...]
-                      },
-                      "edges": {"outgoing": [...], "incoming": [...]},
-                      "related_ids": ["<id>", ...]   # deduped union
-                    }
-
-                Sections not listed in ``include`` are omitted entirely
-                (not emitted as empty keys). ``related_ids`` holds the
-                deduped union of every id referenced across the included
-                sections — sorted for determinism.
-            """
-            logger.info(
-                "lithos_related: called",
-                extra={"id": id, "include": include, "depth": depth, "namespace": namespace},
-            )
-            tracer = get_tracer()
-            with tracer.start_as_current_span("lithos.tool.related") as span:
-                span.set_attribute("lithos.tool", "lithos_related")
-                span.set_attribute("lithos.id", id)
-
-                if not self.knowledge.has_document(id):
-                    return error_envelope("doc_not_found", f"Document not found: {id}")
-
-                requested = include if include is not None else list(_RELATED_INCLUDES)
-                selected = [k for k in _RELATED_INCLUDES if k in requested]
-                span.set_attribute("lithos.include", ",".join(selected))
-
-                depth = min(max(depth, 1), 3)
-                span.set_attribute("lithos.depth", depth)
-
-                result: dict[str, Any] = {
-                    "id": id,
-                    "included": selected,
-                }
-                related_ids: set[str] = set()
-
-                # --- links (NetworkX wiki-links) ---------------------------
-                if "links" in selected:
-                    links = self.graph.get_links(doc_id=id, direction="both", depth=depth)
-                    outgoing = [{"id": ln.id, "title": ln.title} for ln in links.outgoing]
-                    incoming = [{"id": ln.id, "title": ln.title} for ln in links.incoming]
-                    result["links"] = {"outgoing": outgoing, "incoming": incoming}
-                    related_ids.update(ln.id for ln in links.outgoing)
-                    related_ids.update(ln.id for ln in links.incoming)
-
-                # --- provenance (frontmatter derived_from_ids) -------------
-                if "provenance" in selected:
-                    sources = self._bfs_provenance(id, "sources", depth)
-                    derived = self._bfs_provenance(id, "derived", depth)
-                    unresolved_sources = sorted(self.knowledge.get_unresolved_sources(id))
-                    result["provenance"] = {
-                        "sources": sources,
-                        "derived": derived,
-                        "unresolved_sources": unresolved_sources,
-                    }
-                    related_ids.update(s["id"] for s in sources)
-                    related_ids.update(d["id"] for d in derived)
-
-                # --- edges (LCMA edges.db) ---------------------------------
-                if "edges" in selected:
-                    if self._config.lcma.enabled:
-                        # Fan out both directions; caller rarely wants just one.
-                        outgoing_edges = await self.projection.list_edges(
-                            from_id=id, namespace=namespace
-                        )
-                        incoming_edges = await self.projection.list_edges(
-                            to_id=id, namespace=namespace
-                        )
-                    else:
-                        outgoing_edges = []
-                        incoming_edges = []
-                    result["edges"] = {
-                        "outgoing": outgoing_edges,
-                        "incoming": incoming_edges,
-                    }
-                    for edge in outgoing_edges:
-                        tid = edge.get("to_id")
-                        if isinstance(tid, str):
-                            related_ids.add(tid)
-                    for edge in incoming_edges:
-                        fid = edge.get("from_id")
-                        if isinstance(fid, str):
-                            related_ids.add(fid)
-
-                # Exclude the document itself from related_ids so callers
-                # can iterate without filtering.
-                related_ids.discard(id)
-                result["related_ids"] = sorted(related_ids)
-                span.set_attribute("lithos.related_count", len(related_ids))
-
-                return result
-
-        @self.mcp.tool()
-        @tool_metrics()
-        async def lithos_tags(
-            prefix: str | None = None,
-        ) -> dict[str, dict[str, int]]:
-            """Get all tags with document counts.
-
-            Args:
-                prefix: Optional prefix filter (case-insensitive). Only tags starting with this prefix are returned.
-
-            Returns:
-                Dict with tags mapping tag name to count
-            """
-            logger.info("lithos_tags prefix=%s", prefix)
-            tracer = get_tracer()
-            with tracer.start_as_current_span("lithos.tool.tags") as span:
-                span.set_attribute("lithos.tool", "lithos_tags")
-                tags = await self.knowledge.get_all_tags()
-                if prefix is not None:
-                    tags = {k: v for k, v in tags.items() if k.lower().startswith(prefix.lower())}
-                span.set_attribute("lithos.tag_count", len(tags))
-                return {"tags": tags}
 
         # ==================== Coordination Tools ====================
 
@@ -2833,30 +1512,6 @@ class LithosServer:
                 reblocked = await self.coordination.newly_reblocked_by(task_id, prior_status)
                 span.set_attribute("lithos.reblocked_count", len(reblocked))
                 return {"success": True, "reblocked": reblocked}
-
-        @self.mcp.tool()
-        @tool_metrics()
-        async def lithos_node_stats(
-            node_id: str,
-        ) -> dict[str, Any]:
-            """View a note's salience score, retrieval stats, and penalty counts.
-
-            Args:
-                node_id: The document ID to look up stats for
-
-            Returns:
-                Dict with salience, retrieval_count, cited_count, ignored_count,
-                misleading_count, and other stats fields.
-                Returns error envelope if node_id does not match any known document.
-            """
-            logger.info("lithos_node_stats node_id=%s", node_id)
-            tracer = get_tracer()
-            with tracer.start_as_current_span("lithos.tool.node_stats") as span:
-                span.set_attribute("lithos.tool", "lithos_node_stats")
-                span.set_attribute("lithos.node_id", node_id)
-
-                result = await self.memory.node_stats(node_id)
-                return result if isinstance(result, dict) else dataclasses.asdict(result)
 
         @self.mcp.tool()
         @tool_metrics()
